+++ kirlent/docutils/__init__.py
@@ -9,9 +9,14 @@ __version__ = "0.1.0"
 
 from docutils.core import publish_cmdline
 
-from . import html5
+from . import html5, impressjs
 
 
 def publish_cmdline_html5(*args, **kwargs):
     """Run utility for converting an RST file to HTML5."""
     publish_cmdline(*args, **kwargs, writer=html5.Writer())
+
+
+def publish_cmdline_impressjs(*args, **kwargs):
+    """Run utility for converting an RST file to an impress.js presentation."""
+    publish_cmdline(*args, **kwargs, writer=impressjs.Writer())