+++ CHANGES.rst
@@ -2,3 +2,4 @@
 ------------------
 
 - Custom HTML5 writer.
+- impress.js writer.